\name{Chi2SignificantContingencyTable}
\alias{Chi2SignificantContingencyTable}
\title{ Uses Chi2Test to see whether the Independence assumption is valid see also: [table,chi2,p] = crosstab(col1,col2)  }
\arguments{
\item{Xtab(1:l+2,1:c+2)}{a crosstabulation produced for example by ContingencyTableSummary}
}
\value{Pvalue                           the Pvalue of the chi2 test}
\description{
 Uses Chi2Test to see whether the Independence assumption is valid see also: [table,chi2,p] = crosstab(col1,col2)  }
\author{
Michael Thrun
}
