\name{ClassCount}
\alias{ClassCount}
\title{ Class Count}
\description{
Calculates the number of points in each group of data.}
\usage{
ClassCount(Cls)
}
\arguments{
  \item{Cls}{ The clusterNumber of a given data.}
}
% \details{}
\value{
\item{uniqueClasses}{ The unique classes in cls}
\item{countPerClass}{ The number of data points in the corresponding unique class.}
\item{numberOfClasses}{ Just that, the number of classes. }
\item{cls}{ colvector(cls) }
}
% \references{ ~put references to the literature/web site here ~ }
\author{ Felix Pape}
% \note{ ~~further notes~~ }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
