\name{ClassFirst}
\alias{ClassFirst}
\title{ Class First}
\description{
Returns the first data set in each class. First according to the sequence of data.}
\usage{
ClassFirst(data, cls)
}
\arguments{
  \item{data}{ A matrix containing data. Cases as rows, variables as columns.}
  \item{cls}{ The clusterNumber of a given data.}
}
%\details{}
\value{
\item{uniqueClasses}{The unique classes in cls}
\item{firstPerClass}{The first data point in a given data.}
}
% \references{ ~put references to the literature/web site here ~ }
\author{ Raphael Paebst}
% \note{ ~~further notes~~  }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
