\name{ClassMin}
\alias{ClassMin}
\title{ Class Min }
\description{
Calculates the minimum of each group of data.}
\usage{
ClassMin(Data, Cls)
}
\arguments{
  \item{Data}{ A matrix containing data. Cases as rows, variables as columns. }
  \item{Cls}{ The clusterNumber of a given element of data. }
}
% \details{}
\value{
\item{UniqueClasses}{The unique classes in cls. }
\item{MinPerClass}{The minimum of the data points for a given element of uniqueClasses. }
}
% \references{ ~put references to the literature/web site here ~ }
\author{ Raphael Paebst}
% \note{ ~~further notes~~  }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{ }}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
