\name{ClassNaNs}
\alias{ClassNaNs}
\title{ Class NaNs}
\description{
Calculates the number of NaNs in a class.}
\usage{
ClassNaNs(Data, Cls)
}
\arguments{
  \item{Data}{ a matrix with cases as rows and variables as columns}
  \item{Cls}{ The clusterNumber of a given element of data.}
}
% \details{ }
\value{
\item{UniqueClasses}{The unique classes in cls. }
\item{NansPerClass}{The number of NaNs among the data points for a given element of uniqueClasses. } 
}
% \references{ ~put references to the literature/web site here ~ }
\author{ Raphael Paebst }
% \note{ ~~further notes~~  }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{ }}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
