\name{ClassPlotPixMatrix}
\alias{ClassPlotPixMatrix}
\title{ ClassPlotPixMatrix(Data,Cls,LowLim,HiLim,XNames,YNames); plot Data matrix as a pixel coulour picture, sorted by Cls and horizontal division}
\arguments{
\item{Data(1:d,1:n)}{data cases in rows, variables in columns}
\item{Cls(1:d,1)}{Class vector}
\item{LowLim,HiLim}{limits for the color axis}
\item{XNames}{names of the data added at x -axis , Names =0 means no names}
\item{YNames}{names of the data added at y -axis , YNames =0 means no names}
}
\value{List with
}
\description{
 ClassPlotPixMatrix(Data,Cls,Names,LowLim,HiLim); plot Data matrix as a pixel coulour picture, sorted by Cls and horizontal division}

