\name{ClassRange}
\alias{ClassRange}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Class Range}
\description{
short description of function working
}
\usage{
%- all arguments (funktionsaufruf)
ClassRange(Data, Cls)
}
\arguments{
  \item{Data}{matrix containing all points}
  \item{Cls}{vector defining a class for each point in data}
}
\value{
%- description of return values
	\item{UniqueClasses}{sorted factor containing each class identification once}
	\item{RangePerClass}{cbind of minimal values per class per dimension and maximum values per class per dimension}
}
\references{
	\url{www.uni-marburg.de/fb12/datenbionik}
	\email{ultsch@mathematik.uni-marburg.de}
}
\author{ Alfred Ultsch }

