\name{ClassSum}
\alias{ClassSum}
\title{ Class Sum}
\description{
Calculates the sum in each group of data}
\usage{
ClassSum(Data, Cls)
}
\arguments{
  \item{Data}{ Matrix containing data. Cases as rows, variables as columns.}
  \item{Cls}{ A vector containing class identifiers.}
}
% \details{}
\value{
\item{uniqueClasses}{The unique classes in cls.}
\item{sumPerClass}{The sum of data points in an element of uniqueClasses. } 
}
% \references{ ~put references to the literature/web site here ~ }
\author{Michael Thrun}
% \note{ ~~further notes~~ }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
