\name{ClassUniqeAnz}
\alias{ClassUniqeAnz}
\title{ Calulates the number of unique points in each group of the data}
\arguments{
\item{Data[d,n]}{d cases,  n variables}
\item{Cls[d]}{Cls(i) == ClusterNumber of Data[i,:]}
}
\value{UniqueClasses       Number of unique classes in Cls}
\description{
 Calulates the number of unique points in each group of the data}
