\name{ContingencyTableSummary}
\alias{ContingencyTableSummary}
\title{ comfortable crosstabulation of instances in two different classifications including class headers, counts & frequencies uses MATLAB's function crosstab}
\arguments{
\item{RowCls,ColCls}{vector of class identifiers (i.e.integers or NaN's) of the same length}
}
\value{Xtable                        die Kontingenz tabelle erganzt um Klassennamen, Summen und Percentages fuer Zeilen und Spalten}
\description{
 comfortable crosstabulation of instances in two different classifications including class headers, counts & frequencies uses MATLAB's function crosstab}
\author{
MT 07/2015}
