\name{CountClsInstances}
\alias{CountClsInstances}
\title{ Count Cls Instances}
\description{
Counts the number and frequencies of the different instances in cls.}
\usage{
CountClsInstances(cls)
}
\arguments{
  \item{cls}{ A vector containing class identifiers}
}
% \details{}
\value{
\item{classCount}{Number of instances of the different values in cls. }
\item{classFrequency}{Percentages of classCount. }
\item{rowID}{The instances corresponding to classCount. }
}
% \references{ ~put references to the literature/web site here ~ }
\author{ Raphael Paebst}
% \note{ ~~further notes~~ }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
