\name{CountSequences}
\alias{CountSequences}
\title{ Count Sequences}
\description{
Assigns an increasing number to consecutive sequences of the same class label. }
\usage{
CountSequences(cls)
}
\arguments{
  \item{cls}{ A vector of class identifiers.}
}
% \details{}
\value{
\item{uniqueClasses}{The unique classes in cls. }
\item{countPerClass }{The number of data points in a given element of uniqueClasses. }
\item{numberOfClasses}{The number of classes. }
}
% \references{ ~put references to the literature/web site here ~ }
\author{ Raphael Paebst}
% \note{ ~~further notes~~ }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
