\name{dbt.ClassAnalysis-package}
\alias{dbt.ClassAnalysis-package}
\alias{dbt.ClassAnalysis}
\docType{package}
\title{
A package for class analysis.
}
\description{
A package for class analysis. 
}
\details{
\tabular{ll}{
Package: \tab dbt.ClassAnalysis\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-05-11\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
}
\author{Raphael Paebst} %\references{ }
\keyword{ package }
% \seealso{ }
% \examples{}