\name{InnerClassDistances}
\alias{InnerClassDistances}
\title{ Inner Class Distances }
\description{
Calculates the inner class distances.}
\usage{
InnerClassDistances(Data, Cls)
}
\arguments{
  \item{Data}{ A matrix containing the data. Cases as rows, variables as columns. }
  \item{Cls}{ A vector of class identifiers. }
}
% \details{}
\value{
\item{innerDist}{A vector of all inner class distances. }
\item{innerClassNumbers}{The corresponding class numbers. }
}
% \references{ ~put references to the literature/web site here ~ }
\author{ Raphael Paebst }
% \note{ ~~further notes~~ }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
