\name{ IinterClassDistances}
\alias{InterClassDistances}
\title{ Inter Class Distances. }
\description{
Calculates the inter class distances.}
\usage{
InterClassDistances(Data, Cls)
}
\arguments{
  \item{Data}{ A matrix containing data. Cases as rows, variables as columns.}
  \item{Cls}{ A vector defining a class for each point in Data }
}
% \details{}
\value{
\item{interDist}{A vector of all inter class distances. }
\item{classNrA}{One of two corresponding class numbers for the calculation of the inter class distance. }
\item{classNrB}{The other class number needed. }
}
% \references{ ~put references to the literature/web site here ~ }
\author{ Raphael Paebst }
% \note{ ~~further notes~~ }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
