\name{MaxInnerClassDistance}
\alias{MaxInnerClassDistance}
\title{ Maximum inner Class Distance }
\description{
The maximum euklidian distance of any two points in the same class.}
\usage{
MaxInnerClassDistance(data, cls)
}
\arguments{
  \item{data}{ A matrix containing data. Cases as rows, variables as columns. }
  \item{cls}{ A vector of class identifiers. }
}
% \details{}
\value{
\item{maxInnerDist}{The maximum euklidian distance of any two points in the same class. }
}
% \references{ ~put references to the literature/web site here ~ }
\author{ Raphael Paebst }
% \note{ ~~further notes~~ }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
