\name{RenameDescendingClassSize}
\alias{RenameDescendingClassSize}
\title{ RenameDescendingClassSize}
\description{
Renames the classes of a classification in descending order.}
\usage{
RenameDescendingClassSize(cls)
}
\arguments{
  \item{cls}{ A classification. A vector of classes.}
}
% \details{}
\value{
\item{RenamedCls}{The renamed classification. A vector of classes, were the largest class is C1 and so forth}
}
% \references{ ~put references to the literature/web site here ~ }
\author{ Raphael Paebst}
% \note{ ~~further notes~~ }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
