\name{SwitchCls}
\alias{SwitchCls}
\title{ SwitchClass }
\description{
Switches the index of two classes.}
\usage{
SwitchCls(cls, c1, c2)
}
\arguments{
  \item{cls}{ A vector of classifications }
  \item{c1}{ A class number to be exchanged}
  \item{c2}{ The other class number to be exchanged.}
}
% \details{}
\value{
\item{newCls}{A vector of class numbers with c1 changed to c2}  
}
% \references{ ~put references to the literature/web site here ~ }
\author{ Raphael Paebst}
% \note{ ~~further notes~~ }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
