\name{TheSameKey}
\alias{TheSameKey}
\title{ TheSameKey}
\description{
Compares two keys and returns ok, if they are the same. Maybe leave out?}
\usage{
TheSameKey(keyA, keyB)
}
\arguments{
  \item{keyA}{ A key represented by a vector of integers.}
  \item{keyB}{ Another key in the form of a vector of integers.}
}
% \details{}
\value{
\item{KeyOk}{A boolean value. True if both keys are the same, false otherwise.}
}
% \references{ ~put references to the literature/web site here ~ }
\author{ Raphael Paebst}
% \note{ ~~further notes~~ }
% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
% \examples{}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
